requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
python-decouple==3.8
redis==5.0.1
httpx==0.25.2
user-agents==2.2.0
geoip2==4.7.0
python-dateutil==2.8.2
